factor out jolokia collector for fufure JMX type of collector
