+++ collectors/builtin/tomcat.py
@@ -1,13 +1,10 @@
 #!/usr/bin/python
 
-import urllib2
-import json
-from datetime import datetime
-from collectors.lib.collectorbase import CollectorBase
-from collectors.lib.counter_processor import CounterPorcessor
+from collectors.lib.jolokia import JolokiaCollector
+from collectors.lib.jolokia import JolokiaParserBase
 
 
-class Tomcat(CollectorBase):
+class Tomcat(JolokiaCollector):
     JMX_REQUEST_JSON = r'''[
     {
         "type" : "read",
@@ -40,90 +37,21 @@ class Tomcat(CollectorBase):
   ]'''
 
     def __init__(self, config, logger, readq):
-        super(Tomcat, self).__init__(config, logger, readq)
-        self.parsers = {        # key is the mbean name
+        parsers = {        # key is the mbean name
             "Catalina:name=\"http-bio-8080\",type=GlobalRequestProcessor": JolokiaGlobalRequestProcessorParser(
-                self._logger),
-            "java.lang:type=Memory": JolokiaMemoryParser(self._logger),
-            "Catalina:name=\"http-bio-8080\",type=ThreadPool": JolokiaThreadPoolParser(self._logger),
-            "java.lang:type=Threading": JolokiaThreadingParser(self._logger),
-            "java.lang:name=PS Scavenge,type=GarbageCollector": JolokiaGCParser(self._logger),
-            "java.lang:type=OperatingSystem": JolokiaOSParser(self._logger)
+                logger),
+            "java.lang:type=Memory": JolokiaMemoryParser(logger),
+            "Catalina:name=\"http-bio-8080\",type=ThreadPool": JolokiaThreadPoolParser(logger),
+            "java.lang:type=Threading": JolokiaThreadingParser(logger),
+            "java.lang:name=PS Scavenge,type=GarbageCollector": JolokiaGCParser(logger),
+            "java.lang:type=OperatingSystem": JolokiaOSParser(logger)
         }
+        super(Tomcat, self).__init__(config, logger, readq, Tomcat.JMX_REQUEST_JSON, parsers)
 
-    def __call__(self):
-        conn = None
-        try:
-            self.log_info("")
-            protcol = self.get_config("protocol", "http")
-            port = self.get_config("port", "8080")
-            url = "%(protcol)s://localhost:%(port)s/jolokia" % dict(protcol=protcol, port=port)
-            req = urllib2.Request(url, Tomcat.JMX_REQUEST_JSON, {'Content-Type': 'application/json'})
-            conn = urllib2.urlopen(req)
-            status_code = conn.getcode()
-            if status_code != 200:
-                self.log_error("failed to query jolokia endpoint of tomcat. return code %d", status_code)
-                return
-            resp_json = conn.read()
-            json_obj_or_list = json.loads(resp_json)
-            if type(json_obj_or_list) is not list and json_obj_or_list['status'] != 200:
-                self.log_error('failed request. status code %d, error %s', json_obj_or_list['status'],
-                               json_obj_or_list['error'])
-                return
-
-            for json_dict in json_obj_or_list:
-                mbean_key = json_dict["request"]["mbean"]
-                try:
-                    parser = self.parsers[mbean_key]
-                    if parser:
-                        parser.parse(json_dict, self._readq)
-                    else:
-                        self.log_error("failed to instantiate parser %s, skip.", mbean_key)
-                except Exception as e:
-                    self.log_exception("exception when parsing %s. skip", mbean_key)
-        except Exception as e:
-            self.log_exception("unexpected error")
-        finally:
-            if conn is not None:
-                conn.close()
-
-
-class JolokiaParserBase(object):
-    def __init__(self, logger):
-        self.logger = logger
-        self._counter_processors = {}
-
-    def parse(self, json_dict, readq):
-        status = json_dict["status"]
-        if status != 200:
-            raise IOError("status code %d" % status)
-        ts = json_dict["timestamp"]
-        value_dict = self.metric_dict(json_dict)
-        for name in self.valid_metrics():
-            if name in value_dict:
-                val = value_dict[name]
-                if self.iscounter(name):
-                    # origval = val
-                    val = self._get_counter_processor(name).process_counter(ts, val)
-                    # print '%s %s orig=%d, processed=%d' % (datetime.now(), name, origval, val)
-                readq.nput("%s %d %d" % (self.metric_name(name), ts, val))
-
-    def metric_dict(self, json_dict):
-        return json_dict["value"]
-
-    def valid_metrics(self):
-        return []
-
-    def metric_name(self, name):
-        return "%s.%s" % ("tomcat", name)
-
-    def iscounter(self, name):
-        return False
-
-    def _get_counter_processor(self, name):
-        if not self._counter_processors.has_key(name):
-            self._counter_processors[name] = CounterPorcessor()
-        return self._counter_processors[name]
+    def __call__(self, *arg):
+        protocol = self.get_config("protocol", "http")
+        port = self.get_config("port", "8080")
+        super(Tomcat, self).__call__(protocol, port)
 
 
 class JolokiaGlobalRequestProcessorParser(JolokiaParserBase):
